[Maps] Refactor MBMapContainer (#45275)
* [Maps] refactor MBMapContainer

* remove debug code

* remove isDrawingFilter from MBMapContainer connector

* remove setTooltipState from MBMapContainer connector

* add guard around removing map handlers

* only render TooltipControl when mbMap exists, put mbMap on state

* only create DrawControl component when mbMap exists
