+++ x-pack/legacy/plugins/maps/public/components/geometry_filter_form.js
@@ -209,4 +209,3 @@ export class GeometryFilterForm extends Component {
     );
   }
 }
-
