+++ x-pack/legacy/plugins/maps/public/connected_components/map/mb/draw_control/draw_tooltip.js
@@ -9,14 +9,14 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import { EuiText } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import { DRAW_TYPE } from '../../../../common/constants';
+import { DRAW_TYPE } from '../../../../../common/constants';
 import mapboxgl from 'mapbox-gl';
 import { I18nProvider } from '@kbn/i18n/react';
 
 export class DrawTooltip extends Component {
 
-  constructor() {
-    super();
+  constructor(props) {
+    super(props);
     this._tooltipContainer = document.createElement('div');
     this._mbPopup = new mapboxgl.Popup({
       closeButton: false,
