+++ x-pack/legacy/plugins/maps/public/connected_components/map/mb/index.js
@@ -13,17 +13,13 @@ import {
   setMouseCoordinates,
   clearMouseCoordinates,
   clearGoto,
-  setTooltipState,
   setMapInitError,
-  updateDrawState
 } from '../../../actions/map_actions';
 import {
   getTooltipState,
   getLayerList,
   getMapReady,
   getGoto,
-  getDrawState,
-  isDrawingFilter,
   getScrollZoom
 } from '../../../selectors/map_selectors';
 import { getInspectorAdapters } from '../../../reducers/non_serializable_instances';
@@ -35,8 +31,6 @@ function mapStateToProps(state = {}) {
     goto: getGoto(state),
     inspectorAdapters: getInspectorAdapters(state),
     tooltipState: getTooltipState(state),
-    drawState: getDrawState(state),
-    isDrawingFilter: isDrawingFilter(state),
     scrollZoom: getScrollZoom(state)
   };
 }
@@ -63,15 +57,9 @@ function mapDispatchToProps(dispatch) {
     clearGoto: () => {
       dispatch(clearGoto());
     },
-    setTooltipState(tooltipState) {
-      dispatch(setTooltipState(tooltipState));
-    },
     setMapInitError(errorMessage) {
       dispatch(setMapInitError(errorMessage));
     },
-    disableDrawState() {
-      dispatch(updateDrawState(null));
-    }
   };
 }
